refactor(nodes): migrate tool_vectordb to @tool_function pattern per PR #599
Remove VectorDBDriver(ToolsBase) and move search/upsert/delete tool logic
to @tool_function decorated methods on IInstance. Store config and backend
connection directly on IGlobal instead of through a driver object.
